gated_delta_update VJP: address review findings
- gated_delta_vjp / gated_delta_vjp_metal: initialize default state as
  fp32 to match the existing gated_delta_update path. Previously the VJP
  modules used q.dtype, silently downgrading the recurrent state to
  bf16 during Qwen3.5 / Qwen3-Next training when state=None.

- gated_delta_update: training=True now selects the Metal VJP only when
  GPU/Metal is available, mask is None, Dk%32==0 and Dv%4==0. Otherwise
  it falls back to the Python VJP, which has no shape constraints and
  also runs on CPU. Previously training=True unconditionally invoked
  the Metal kernel and crashed for non-GPU runs and for shapes the
  kernel does not handle.

- tests/test_models.py: add test_gated_delta_vjp_metal_gradients_match_python
  to exercise mx.grad through both the Metal and Python VJP and compare
  gradients for q, k, v, a, b, A_log, dt_bias, state. The previous
  Metal-vs-Python test only compared forward outputs, so a broken Metal
  backward could have passed.

+++ mlx_lm/models/gated_delta_vjp.py
@@ -147,7 +147,7 @@ def gated_delta_update_vjp(
         k = mx.repeat(k, repeat_factor, axis=-2)
 
     if state is None:
-        state = mx.zeros((B, Hv, Dv, Dk), dtype=q.dtype)
+        state = mx.zeros((B, Hv, Dv, Dk), dtype=mx.float32)
 
     # Chunked forward; each chunk is a pure function of the incoming state,
     # so autodiff propagates gradients correctly across the recurrence.
